rename: 기본프로필 이미지 파일명 변경
